miner: Move set-addrs command to a new 'actor' namespace
